refactor: extract modules from computation_engine.py (issue #146)
Extract focused responsibilities into computation_engine_lib/ helpers:
- change_tracker.py: ForecastChangeTracker
- price_calculator.py: effective cheap price + solar-weighted FIT
- mode_decision.py: active mode + decision-log maintenance
- spike_analyzer.py: conservative spike analysis + reserve SOC
- excess_solar_signals.py: excess-solar/load-shift signals
- forecast_accuracy.py: planned-vs-actual forecast comparisons
- weather_diagnostics.py: weather-learning diagnostic population

ComputationEngine retains orchestration and lifecycle concerns.
Behavior unchanged; delegation pattern preserves API compatibility.

Closes #146
